Single-player agent/requirements.txt
neat-python==0.92
numpy==1.19.0
Pillow==8.3.2
pygame==1.9.6
six==1.15.0
